+++ docs/en/changes/changes.md
@@ -14,6 +14,7 @@
 * Add `LatestLabeledFunction` for meter.
 * MAL Labeled metrics support additional attributes.
 * Bump up netty to 4.2.9.Final.
+* Add support for OpenSearch/ElasticSearch client certificate authentication.
 
 #### UI
 * Fix the missing icon in new native trace view.